feat(user): finish all user logics
