--- src/@types/ali-oss/index.d.ts
@@ -1,5 +1,0 @@
-declare module "ali-oss" {
-  import OSS from "ali-oss";
-  export { STS };
-  export default OSS;
-}
