+++ src/@types/express/index.d.ts
@@ -1,10 +1,9 @@
-import { UserModel } from "../../models/user";
+import { UserInfo } from "../../middlewares/authenticate";
 
 declare module "express-serve-static-core" {
-  interface Request {
-    auth: {
-      user: UserModel;
-      selfCheckRequired?: boolean;
-    };
+    interface Request {
+      auth: {
+        user: UserInfo;
+      };
+    }
   }
-}
