+++ src/middlewares/authenticate.ts
@@ -1,6 +1,8 @@
 import { NextFunction, Request, Response } from "express";
 import jwt from "jsonwebtoken";
-import User from "../models/user";
+import { gql } from "graphql-request";
+import { client } from "..";
+
 export interface JwtPayload {
   _id: string;
   email: string;
@@ -26,6 +28,21 @@ export interface JwtVerifyPayload {
   phone: string;
   code: string; // hash加密后的验证码
 }
+export interface UserInfo {
+  uuid: string;
+  username: string;
+  password: string;
+  role: string;
+  realname: string;
+  email: string;
+  phone: string;
+  student_no: string;
+  department: string;
+  class: string;
+  tsinghua_email: string;
+  github_id: string;
+  id: string;
+}
 
 /**
  * Middleware: validate user authorizations; reject if necessary
@@ -42,40 +59,53 @@ const authenticate: (
     }
 
     const token = authHeader.substring(7);
-    return jwt.verify(token, process.env.SECRET!, (err, decoded) => {
+    return jwt.verify(token, process.env.SECRET!, async (err, decoded) => {
       if (err || !decoded) {
         return res
           .status(401)
           .send("401 Unauthorized: Token expired or invalid");
       }
-
-      const payload = decoded as JwtPayload;
-
-      User.findOne({ email: payload.email }, (error: any, user: any) => {
-        if (error) {
-          console.error(err);
-          return res.status(500).end();
-        }
+      const payload = decoded as JwtUserPayload;
+      console.log(payload.uuid); //delete
+      try {
+        const user = (await client.request(
+          gql`
+            query MyQuery($uuid: uuid!) {
+              users(where: {uuid: {_eq: $uuid}}) {
+                uuid,
+                username,
+                password,
+                role,
+                realname,
+                email,
+                phone,
+                student_no,
+                department,
+                class,
+                tsinghua_email,
+                github_id,
+                id
+              }
+            }
+          `,
+          {
+            uuid: payload.uuid
+          }
+        )).users[0];
         if (!user) {
           return res.status(401).send("401 Unauthorized: Permission denied");
         }
         req.auth = { user };
-        // use authenticate() to accept all registered users
-        if (!acceptableRoles || acceptableRoles.length === 0) {
+        if (!acceptableRoles || acceptableRoles.length === 0 || acceptableRoles.includes(user.role)) {
           return next();
         }
-        if (!acceptableRoles.includes(user.role)) {
-          // leave it to next() to see if it indeed accesses `self`
-          if (acceptableRoles.includes("self")) {
-            req.auth.selfCheckRequired = true;
-            return next();
-          }
-
+        else {
           return res.status(401).send("401 Unauthorized: Permission denied");
         }
-
-        next();
-      });
+      } catch (err) {
+        console.error(err);
+        return res.status(500).send(err);
+      }
     });
   };
 };
